requirements.txt
trimesh==4.5.0
numpy==1.26.0
ffmpeg-python==0.2.0
Pillow==10.4.0
tqdm==4.66.0
